fix: update network body max size handling and caching
- Changed the return type of `getNetworkBodyMaxSize` from `int?` to `double?` in the API and updated related method calls accordingly.
- Renamed `registerW3CFlagsListener` to `registerFeatureFlagsListener` for consistency across the codebase.
- Introduced a callback mechanism in `FeatureFlagsManager` to handle changes in network body max size, ensuring proper cache management in `NetworkManager`.
- Updated tests to reflect the new method names and ensure functionality remains intact.

+++ android/src/main/java/com/instabug/flutter/modules/InstabugApi.java
@@ -516,13 +516,13 @@ public class InstabugApi implements InstabugPigeon.InstabugHostApi {
     }
 
     @Override
-    public void getNetworkBodyMaxSize(@NonNull InstabugPigeon.Result<Long> result) {
+    public void getNetworkBodyMaxSize(@NonNull InstabugPigeon.Result<Double> result) {
         ThreadManager.runOnMainThread(
             new Runnable() {
                 @Override
                 public void run() {
                     try {
-                        long networkCharLimit = InternalCore.INSTANCE.get_networkLogCharLimit();
+                        double networkCharLimit = InternalCore.INSTANCE.get_networkLogCharLimit();
                         result.success(networkCharLimit);
                     } catch (Exception e) {
                         e.printStackTrace();
